Fixed icon changing

+++ app/src/main/java/ru/vtosters/lite/utils/Preferences.java
@@ -337,7 +337,7 @@ public class Preferences {
     }
 
     public static long VKVersionBuildLong() {
-        return (long) VKVersionBuild();
+        return VKVersionBuild();
     }
 
     public static String getLocale() {
